+++ question-answering-bert-qat/onnxovep_optimum_inference/bert_inference_optimum_ort_ovep.py
@@ -88,8 +88,10 @@ def run_inference():
         rows.append([context,question,answer])
     headers = ['Context','Question', 'Answer']
     answers = pd.DataFrame(rows, columns=headers)
-    answers.to_csv(args.outputpath, index=False)
-    print('Results is stored in Output CSV file')
-
+    if os.path.exists(os.path.dirname(args.outputpath)) and pathlib.Path(args.outputpath).suffix =='.csv':
+        answers.to_csv(args.outputpath, index=False)
+        print('Results is stored in Output CSV file')
+    else:
+        print('Output is not stored in Output CSV file as ouputpath not exists')
 if __name__ == '__main__':
     run_inference()
